@@ -1,5 +1,5 @@
 name: f1_pet_project
-description: F1 pet project. State Management - Elementary. Navigation - Auto Route. API - http://ergast.com/mrd/.
+description: F1 pet project. State Management - Elementary. Navigation - Auto Route. API - http://api.jolpi.ca/ergast/f1/.
 
 publish_to: "none"
 
@@ -15,58 +15,65 @@ publish_to: "none"
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 # In Windows, build-name is used as the major, minor, and patch parts
 # of the product and file versions while build-number is used as the build suffix.
-version: 1.0.0+1
+version: 1.0.0+2
 
 environment:
-  sdk: ">=3.0.0"
+  sdk: ">=3.5.0 <4.0.0"
+  flutter: ">=3.35.3"
 
 dependencies:
-  auto_route: ^7.8.4
+  auto_route: ^10.1.2
   colorful_safe_area: ^1.0.0
   cookie_jar: ^4.0.8
-  dartdoc: ^6.3.0 # do not update because of - analyzer from dark sdk
-  dio: ^5.4.0
-  dio_cookie_manager: ^3.1.1
-  elementary: ^3.0.0
-  elementary_helper: ^1.0.1
-  extended_image: ^9.1.0
+  dartdoc: ^8.3.4
+  dio: ^5.9.0
+  dio_cookie_manager: ^3.3.0
+  elementary: ^3.2.1
+  elementary_helper: ^1.0.3
+  extended_image: ^10.0.1
   flutter:
     sdk: flutter
-  flutter_compass: ^0.8.0
+  flutter_compass: ^0.8.1
   flutter_html: 3.0.0-alpha.5
   flutter_localizations:
     sdk: flutter
-  fluttertoast: ^8.2.4
-  geolocator: ^10.1.1
-  horizontal_data_table: ^4.3.1
-  image: ^4.1.7
-  intl: 0.19.0
-  json_annotation: ^4.8.1
-  json_serializable: ^6.7.1
-  loading_animation_widget: ^1.2.0+4
-  package_info_plus: ^8.3.0
-  path_provider: ^2.1.2
-  permission_handler: ^11.3.0
-  provider: ^6.1.1
-  quiver: ^3.2.1
-  responsive_framework: ^1.1.1
+  flutter_local_notifications: ^19.4.2
+  fluttertoast: ^8.2.14
+  geolocator: ^14.0.1
+  horizontal_data_table: ^4.3.3
+  image: ^4.5.4
+  intl: ^0.20.2
+  json_annotation: ^4.9.0
+  json_serializable: ^6.11.1
+  loading_animation_widget: ^1.3.0
+  package_info_plus: ^9.0.0
+  path_provider: ^2.1.5
+  permission_handler: ^12.0.1
+  provider: ^6.1.5
+  quiver: ^3.2.2
+  responsive_framework: ^1.5.1
   scroll_to_index: ^3.0.1
-  table_calendar: ^3.0.9
-  url_launcher: ^6.2.4
+  table_calendar: ^3.2.0
+  url_launcher: ^6.3.2
   visibility_detector: ^0.4.0+2
-  yandex_mapkit: ^3.4.0
-  flutter_local_notifications: ^19.1.0
+  yandex_mapkit:
+    git:
+      url: https://github.com/DaniilPavlov/yandex_mapkit.git
+      ref: 434784ed3a2aa70d656c8b2d9924ddc31b18ce15
 
 dev_dependencies:
-  auto_route_generator: ^7.3.2
-  build_runner: ^2.4.8
-  elementary_test: ^2.0.0
+  auto_route_generator: ^10.2.4
+  build_runner: ^2.7.1
+  elementary_test: ^2.1.1
   # flutter_native_splash: ^2.2.17
   # flutter_launcher_icons: ^0.9.2
   flutter_test:
     sdk: flutter
-  mocktail: ^1.0.3
-  flutter_lints: ^3.0.1
+  mocktail: ^1.0.4
+  meykuless_lint_rules:
+    git:
+      url: https://github.com/DaniilPavlov/meykuless_lint_rules.git
+      ref: 3a306a7d92684ba4402de5df59f30c5e49f1fff8
 
 # flutter_icons:
 #   android: "ic_launcher"
