@@ -29,7 +29,7 @@ class BlackButton extends StatelessWidget {
         padding: const EdgeInsets.symmetric(vertical: 12, horizontal: 17),
         width: MediaQuery.of(context).size.width,
         decoration: BoxDecoration(
-          color: isDisabled ? AppTheme.black.withOpacity(0.3) : AppTheme.black,
+          color: isDisabled ? AppTheme.black.withValues(alpha: 0.3) : AppTheme.black,
           borderRadius: BorderRadius.circular(50),
           boxShadow: haveShadow && !isDisabled
               ? [
@@ -54,7 +54,7 @@ class BlackButton extends StatelessWidget {
                 text,
                 style: AppStyles.h3.copyWith(
                   color:
-                      isDisabled ? AppTheme.red.withOpacity(0.3) : AppTheme.red,
+                      isDisabled ? AppTheme.red.withValues(alpha: 0.3) : AppTheme.red,
                 ),
               )
             else
