@@ -65,7 +65,7 @@ class MapControlsWidget extends StatelessWidget {
               child: CustomIconButton(
                 image: 'assets/icons/location_user.png',
                 onPressed: onUserLocationPressed,
-                backgroundColor: AppTheme.red.withOpacity(0.75),
+                backgroundColor: AppTheme.red.withValues(alpha: 0.75),
                 iconColor: Colors.white,
               ),
             ),
