@@ -2,5 +2,5 @@ class StaticData {
   static const defaultHorizontalPadding = 12.0;
   static const defaultVerticallPadding = 20.0;
 
-  static const apiUrl = 'http://ergast.com/api/f1/';
+  static const apiUrl = 'http://api.jolpi.ca/ergast/f1/';
 }
